Fix data directory ownership
When using a volume mount for the Rainloop data directory, ownership on
Linux seems to not match that of the www user set up in the container.
This forces ownership to www prior to running the services. However,
this also changes ownership on the host side, forcing a sudo to
recuperate the files. Other images out there seem to use a UID/GID
parameterization of the image, which I dislike. Must think.